Remove old components
