+++ src/custom/components/Header/NetworkSelector/NetworkSelectorMod.tsx
@@ -1,3 +1,5 @@
+import { ArrowDownCircle } from 'react-feather'
+import styled from 'styled-components/macro'
 import { Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
 import { getWeb3ReactConnection } from '@cow/modules/wallet/web3-react/connection'
@@ -13,19 +15,12 @@ import { useCloseModal, useModalIsOpen, useOpenModal, useToggleModal } from 'sta
 import { /*addPopup,*/ ApplicationModal } from 'state/application/reducer'
 import { updateConnectionError } from 'state/connection/reducer'
 import { useAppDispatch } from 'state/hooks'
-import styled from 'styled-components/macro'
 import { ExternalLink, MEDIA_WIDTHS } from 'theme'
 import { replaceURLParam } from 'utils/routes'
 import { isChainAllowed, switchChain } from 'utils/switchChain'
 // import { isMobile } from 'utils/userAgent'
 
 // Mod imports
-import {
-  ActiveRowLinkList,
-  ActiveRowWrapper,
-  FlyoutHeader,
-  LinkOutCircle,
-} from '@src/components/Header/NetworkSelector'
 import { transparentize, darken } from 'polished'
 import { getExplorerBaseUrl } from 'utils/explorer'
 import { SUPPORTED_CHAIN_IDS, supportedChainId } from 'utils/supportedChainId'
@@ -35,7 +30,8 @@ import { useRemovePopup, useAddPopup } from 'state/application/hooks'
 import { useTradeTypeInfo } from '@cow/modules/trade'
 import { useMediaQuery, upToMedium } from 'hooks/useMediaQuery'
 
-/* const ActiveRowLinkList = styled.div`
+
+export const ActiveRowLinkList = styled.div`
   display: flex;
   flex-direction: column;
   padding: 0 8px;
@@ -66,7 +62,13 @@ export const ActiveRowWrapper = styled.div`
 export const FlyoutHeader = styled.div`
   color: ${({ theme }) => theme.text1};
   font-weight: 400;
-` */
+`
+
+export const LinkOutCircle = styled(ArrowDownCircle)`
+  transform: rotate(230deg);
+  width: 16px;
+  height: 16px;
+`
 export const FlyoutMenu = styled.div`
   position: absolute;
   top: 54px;
